Use a normal comparator in the RangeMap
Using the std::greater_than comparator for the RangeMap lead to
frequent bugs, like trying to use ++ instead of -- to iterate
forward through the map, or comparing the of the -- iterator
to end() instead of begin().

Stop using the greater_than comparator, and instead write a
function that produces the same iterator result that lower_bound()
did when used with the greater_than comparator.