Improve performance of DWARF parsing
Parsing a single executable with a hundreds of MB of debug data
could take over 30 minutes.  This is entirely unnecessary as in
a profiling run on such a large executable is likely to touch
only a small minority of its compile units.

Optimize the DWARF parsing by processing individual compile units
lazily.  Unfortunately, libdwarf doesn't make this as easy as it
really should be.  libdwarf *requires* that we iterate over all
compile units to locate the one we're interested in rather than
caching a Dwarf_Die structure or even re-loading the DIE from
a cached offset.

This brings the time to look up a single address in a file with
400MB of debug data from 30 minutes to under 30 seconds.