Handle exec() events from processes
Previously if a processed called exec after pmcstat sample
collection had started, the samples against the new executable
would be counted against the old executable and produce garbage
profiling data.  Unfortunately this was trivial to reproduce by
using pmcstat to execute a particular executable in a profiling
session.

Fix this by ensuring that when exec() events are received, we
always create a new Process() object for that pid.