Further C++14-ization: delete copy constructors
Delete copy constructors from classes that were not intended to be
copied.  This turned up a couple instances where copies were
inadvertently being used, so fix those to be references.  This also
requires implementing move constructors in a couple of places so
that we can use move semantics rather than copies of temporary
objects.