Handle subprograms not sorted by low_pc
DWARF does not guarantee that subprogram DIEs will be sorted by
their address, which we were assuming to be true.  Fix this by
instead collecting all subroutine DIEs for a CU in a map that is
sorted by address.  When mapping frames, we just have to look up
the frame offset in the map to find the subroutine DIE for it, then
map all frames that fall within that subroutine.