Handle compile units that are not sorted by lowpc
Like other DIEs, DIEs representing compile unit are not guaranteed
to be sorted.  Fix this by iterating over the top-level DIEs in
a compile unit and caching the DIEs in a sorted data structure
suitable for quick look-ups,

This has required refactoring the DwarfCompileUnit class into one
suitable for iterating over compile units in the DWARF data and
an immutable class that caches information about a given compile
unit.  A third class, DwarfCompileUnitDie, has been introduced to
represent a DIE with a tag of compile_unit (this is different from
a compile unit.  Thanks, DWARF!)

The DwarfDieLookup class has been refactored to be a general
structure suitable for looking up items that exist at non-
overlapping address ranges, and renamed to DwarfRangeLookup,

Unfortunately, this version only works with the GPL'ed version of
libdwarf.  The elftoolchains version does not allow iterating over
a DIE that is not part of the current compile unit.  I will
investigate fixing this restriction in the near future.