Simplify mocking of global functions
Replace the current brittle method of mocking global functions
with something that is hopefully easier to use.  Rather than
requiring explicit initialization and deinitialization in the
test suite's SetUp() and TearDown(), gate access to the global
mock objects through something test cases can access on the stack.
This ensures that any tests that want to consume the global mocks
have to initialize them to get access to them.