Remove SharedString interning
String interning was introduced for the old DwarfLookup
implementation, which was less efficient in how allocated string
data from libdwarf.  The new DwarfResolver implementation doesn't
seem to be as stupid about memory allocation and hence the interning
winds up being a wash (the memory saved is equal to the memory
consumed to track the interning).

Remove the interning, as this simplifies the code and it's about
3% faster to not do the intern lookup every time we create a
new SharedString.